+++ react/Reviews.tsx
@@ -21,6 +21,7 @@ import AppSettings from '../graphql/appSettings.graphql'
 import ReviewsByProductId from '../graphql/reviewsByProductId.graphql'
 import AverageRatingByProductId from '../graphql/averageRatingByProductId.graphql'
 import ReviewsGraph from './ReviewsGraph'
+import { getBaseUrl } from './utils/baseUrl'
 
 interface Review {
   approved: boolean
@@ -359,7 +360,7 @@ function Reviews() {
   const intl = useIntl()
   const handles = useCssHandles(CSS_HANDLES)
   const { product }: any = useProduct() ?? {}
-  const { productId, productName, link } = product ?? {}
+  const { productId, productName, linkText } = product ?? {}
 
   const [state, dispatch] = useReducer(reducer, initialState)
 
@@ -576,6 +577,8 @@ function Reviews() {
     state.settings,
   ])
 
+  const baseUrl = getBaseUrl()
+
   return (
     <div
       className={`${handles.container} review mw8 center ph5`}
@@ -692,7 +695,7 @@ function Reviews() {
                       {JSON.stringify({
                         '@context': 'http://schema.org',
                         '@type': 'Product',
-                        '@id': link,
+                        '@id': `${baseUrl}/${linkText}/p`,
                         review: {
                           '@type': 'Review',
                           reviewRating: {