+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Use product's URI as @id on `Product` schema for `RatingSummary` and `Reviews` blocks.
 
 ## [3.4.1] - 2022-01-04